// src/interfaces/dashboard/malmatta-nodni/MalmattaNodni.types.ts
export interface MalmattaRecord {
  id: number;
  anu_kramank: string;
  malmatta_number: string;
  ward_kramnak: string;
  plot_number: string;
  khasara_number: string;
  survey_number: string;
  ghar_malkache_nav: string;
  bhogavat_darache_nav: string;
  patta_nagar_layout_society: string;
  kayamcha_patta: string;
  gruhkar_v_bhumikar: string;
  chalu_kar: string;
  ekun_kar_bharne: string;
}